huge reformat & maybe fix toolchain build

+++ tools/build-i386-elf.sh
@@ -8,12 +8,11 @@ reset="\033[0m"
 binv=2.40
 gccv=13.2.0
 target=i386-elf
-prefix="$(pwd)/bin/bin"
+prefix="$(pwd)/bin"
+prefix=$(realpath "$prefix")
 
-# put to path temporary
 export PATH="$prefix:$PATH"
 
-# check if already exists
 if command -v $target-gcc >/dev/null 2>&1 && command -v $target-ld >/dev/null 2>&1; then
 	echo -e "${green}toolchain already built${reset}"
 	echo -e "${yellow}path is set, run make in shiggy${reset}"
@@ -26,16 +25,21 @@ mkdir -p bin
 cd toolchain
 
 echo -e "${cyan}install deps${reset}"
-sudo apt install wget build-essential bison flex libgmp3-dev libmpc-dev libmpfr-dev texinfo -y
+if command -v apt >/dev/null 2>&1; then 
+	sudo apt update
+	sudo apt install wget build-essential bison flex libgmp-dev libmpc-dev libmpfr-dev texinfo nasm qemu-system -y
+elif command -v pacman >/dev/null 2>&1; then
+	sudo pacman -S --needed wget base-devel bison flex gmp libmpc mpfr texinfo nasm qemu-full --noconfirm
+else
+	echo -e "${yellow}unsupported distro, install deps manually${reset}"
+fi
 
 echo -e "${cyan}download src${reset}"
 wget https://ftp.wayne.edu/gnu/binutils/binutils-$binv.tar.gz
 wget https://ftp.wayne.edu/gnu/gcc/gcc-$gccv/gcc-$gccv.tar.gz
 
 echo -e "${cyan}extract src${reset}"
-echo -e "${green}extracting binutils${reset}"
 tar -xvzf binutils-$binv.tar.gz
-echo -e "${green}extracting gcc${reset}"
 tar -xvzf gcc-$gccv.tar.gz
 
 echo -e "${cyan}build binutils${reset}"
@@ -56,11 +60,17 @@ mkdir -p build-gcc
 cd build-gcc
 ../gcc-$gccv/configure --target=$target --prefix=$prefix --disable-nls --enable-languages=c --without-headers
 make all-gcc -j$(nproc) V=1
-make install-gcc
+make install -j$(nproc)
 cd ..
-rm -rf toolchain
 
+if [ ! -f "$prefix/bin/i386-elf-gcc" ]; then
+	echo -e "${yellow}i386-elf-gcc not found in $prefix/bin${reset}"
+	ls -l "$prefix/bin"
+	exit 1
+fi
+
+rm -rf toolchain
 
 echo
 echo -e "${green}toolchain built${reset}"
-echo -e "${yellow}path is set now, run make in shiggy${reset}"
+echo -e "${yellow}run make in shiggy${reset}"
